Moving to dynamic versioning
