+++ pyproject.toml
@@ -1,10 +1,13 @@
 [build-system]
-requires = ["setuptools >= 61.0"]
+requires = [
+  "setuptools >= 61.0",
+  "setuptools_scm[toml] >= 8",
+]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "NREL-sup3r"
-version = "0.1.2"
+dynamic = ["version"]
 description = "Super Resolving Renewable Resource Data (sup3r)"
 keywords = ["sup3r", "NREL"]
 readme = "README.rst"
@@ -61,3 +64,6 @@ repository = "https://github.com/NREL/sup3r"
 [tool.setuptools]
 include-package-data = true
 packages = ["sup3r"]
+
+[tool.setuptools_scm]
+version_file = "sup3r/_version.py"
