+++ sup3r/bias/bias_calc_cli.py
@@ -8,10 +8,10 @@ import os
 
 import click
 
+from sup3r import __version__
 import sup3r.bias.bias_calc
 from sup3r.utilities import ModuleName
 from sup3r.utilities.cli import AVAILABLE_HARDWARE_OPTIONS, BaseCLI
-from sup3r.version import __version__
 
 logger = logging.getLogger(__name__)
