+++ sup3r/pipeline/forward_pass_cli.py
@@ -9,7 +9,7 @@ from inspect import signature
 import os
 
 from sup3r.utilities import ModuleName
-from sup3r.version import __version__
+from sup3r import __version__
 from sup3r.pipeline.forward_pass import ForwardPassStrategy, ForwardPass
 from sup3r.utilities.cli import AVAILABLE_HARDWARE_OPTIONS, BaseCLI
