+++ sup3r/postprocessing/data_collect_cli.py
@@ -5,11 +5,11 @@ import logging
 
 import click
 
+from sup3r import __version__
 from sup3r.postprocessing.collection import CollectorH5, CollectorNC
 from sup3r.utilities import ModuleName
 from sup3r.utilities.cli import AVAILABLE_HARDWARE_OPTIONS, BaseCLI
 from sup3r.utilities.utilities import get_source_type
-from sup3r.version import __version__
 
 logger = logging.getLogger(__name__)
