+++ sup3r/qa/visual_qa_cli.py
@@ -5,8 +5,8 @@ sup3r visual QA module CLI entry points.
 import click
 import logging
 
+from sup3r import __version__
 from sup3r.utilities import ModuleName
-from sup3r.version import __version__
 from sup3r.qa.visual_qa import Sup3rVisualQa
 from sup3r.utilities.cli import BaseCLI
